Check if Flatery is installed

+++ scripts/install/desktop/theme.sh
@@ -5,6 +5,9 @@ sudo pacman -S --needed --noconfirm adwaita-qt5 adwaita-qt6
 sudo pacman -S --needed --noconfirm breeze-gtk
 
 # Install icons
-git clone https://github.com/cbrnix/Flatery.git /tmp/Flatery
-sudo cp -rf /tmp/Flatery/Flatery-* /usr/share/icons/
-rm -rf /tmp/Flatery
+if [ ! -d "/usr/share/icons/Flatery" ]
+then
+    git clone https://github.com/cbrnix/Flatery.git /tmp/Flatery
+    sudo cp -rf /tmp/Flatery/Flatery-* /usr/share/icons/
+    rm -rf /tmp/Flatery
+fi
